Greek support, modular weather, command history for console window
Former-commit-id: b72d0b018c084e80cf4fef77e1b1a81101d6daea
Former-commit-id: 32da6a2998826de6519a901dcff7bf058f689b2f

+++ src/net/torvald/terrarum/realestate/RealEstateCodex.kt
@@ -12,9 +12,9 @@ object RealEstateCodex {
      *
      * Note that a tile can have only ONE owner (as an Actor or Faction ID)
      */
-    private var ownershipRegistry: HashMap<Long, Long> = HashMap()
+    private var ownershipRegistry: HashMap<Long, Int> = HashMap()
 
-    fun setOwner(tileX: Int, tileY: Int, refID: Long) {
+    fun setOwner(tileX: Int, tileY: Int, refID: Int) {
         ownershipRegistry[RealEstateUtility.getAbsoluteTileNumber(tileX, tileY)] = refID
     }
 
@@ -22,6 +22,6 @@ object RealEstateCodex {
         ownershipRegistry.remove(RealEstateUtility.getAbsoluteTileNumber(tileX, tileY))
     }
 
-    fun getOwner(tileX: Int, tileY: Int): Long? =
+    fun getOwner(tileX: Int, tileY: Int): Int? =
             ownershipRegistry[RealEstateUtility.getAbsoluteTileNumber(tileX, tileY)]
 }
